requirements.txt
pandas==1.4.1
Pillow==9.2.0
requests==2.27.1
streamlit==1.5.1
streamlit_lottie==0.0.3
